int to long

+++ src/main/java/com/Ace/sem/Country.java
@@ -7,7 +7,7 @@ package com.Ace.sem;
 public class Country {
     // Declare Variables related to Country Information and Capital City Names
 
-    private int worldPop;
+    private long worldPop;
     private String Code;
     private String Name;
     private String Continent;
@@ -16,11 +16,11 @@ public class Country {
     private String Capital;
 
     // Use Encapsulation : Getter & Setter Method for each variable initialized in above
-    public int getWorldPop() {
+    public long getWorldPop() {
         return worldPop;
     }
 
-    public void setWorldPop(int worldPop) {
+    public void setWorldPop(long worldPop) {
         this.worldPop = worldPop;
     }
